requirements.txt
et-xmlfile==2.0.0
openpyxl==3.1.5
pillow==11.0.0
PyQt6==6.7.1
PyQt6-Qt6==6.7.3
PyQt6-sip==13.8.0
